Features/assume role (#2)
* Implemented an option to be able to assume a distinct role on all terraform command. Currently, the current user must have access to state files in order to execute terragrunt command. The assume role as implemented by terraform only allow the user to assume role on AWS resources operations, but the current user still have to have access to the S3 bucket and the DynamoDB used for locking to be able to use terraform remote state.

This feature allows user to define an assume_role in the terragrunt configuration. Then, all external command will be executed using that role instead of the user role.

This also solve a problem when using switch role in .aws/config. This configuration were not recognized by terraform and it complained that no valid credentials were available.

* Add a brief documentation.

* Fix a typo