Properly handle multiple blocks with same name in a single file
If we declare a block with the same name, the normal behavior is to overwrite the previous definition with the latest one. That works if blocks are defined in separated files. But it did not work properly when blocks were declared in the same file.

I fixed this in the generic merge function, but that only worked if multiple blocks were declared in an included file but not for the main configuration file. So, I forced the main file to also go through the merge process (starting with an empty structure).

I also discovered a test `no-tf-variable` that was working even if it used the same block many times in the configuration. Fixing this problem caused that specific test to fail.

To ensure that we don't break existing configuration without being noticed, I added a warning when a block is overridden by another block with the same name in a file.