Fix Helper functions (#174)
* Fix Helper functions

The implementation of helper functions supported anonymous arguments to simplify the implementation. But, that could lead to unexpected behavior since a user can invoke a function that is not supposed to accept arguments without complaining.

ex. findInParentFolders("..", "test")

Will return a valid value, but probably not the value expected by the caller.

Also, with the new implementation of set_global_variable, it was no longer possible to supply values other than string as the second argument while the goal of this function is to support any value type. Since it is not supported by hcl2, we simply ignore this function for the hcl implementation, but we left it as it was for gotemplate.

* Remove warnings and unused stuff

* Change following comment

* Use the right return type on helper functions