Complete the support of extended configuration files
Add support for various terragrunt configuration files (.hcl.json, .yaml, .config) to let users organize their repository as they want. It is now also possible to prefix the terragrunt configuration file with a dot to hide it and put it on the top of the IDE (separated from the terraform files)
Add a validation to determine if the folder pointed by source is containing terraform files (the previous version only validated this when no source was specified). It has been added as an option to not break existing code or tests, but this new behaviour is the default from the command line.
The cleanup now deletes all files that are containing terraform code (previously, only .tf files were deleted and .tf.json files were not). To achieve this, we no longer use the zglob package and rely on the utils.FindFiles with recursive option enabled.
Moved the config/FindConfigFilesInPath to options package where it is more convenient.
Added a function ConfigPath to find any valid configuration file within a path.
If a user specifies a --terragrunt-config argument, the name specified will also be used to resolve the stack (previously, that name was ignored for that command)
Added a detailed test to validate the various configurations options (including different configuration file format). The test also validates the various output of get-stack.
It is now possible to add a dependency to a parent folder even if that folder does not contain a terragrunt configuration file. Then, all terragrunt folders under that folder will be considered as a dependency.
Add the related documentation.
Update terraform to 0.12.26