Fix problem with custom terragrunt config path (#181) (#182)
* Fix problem with custom terragrunt config path

If a user specified a custom terragrunt config path either by specifying terragrunt-config or defining TERRAGRUNT_CONFIG, the -all and stack commands were not working properly.

Also:
- Handle root folders properly to avoid having to add a terragrunt.ignore file in the folder
- Update terraform to 0.12.25 in build scripts
- Remove some warnings

* Remove warnings

* Fix Helper functions

The implementation of helper functions supported anonymous arguments to simplify the implementation. But, that could lead to unexpected behavior since a user can invoke a function that is not supposed to accept arguments without complaining.

ex. findInParentFolders("..", "test")

Will return a valid value, but probably not the value expected by the caller.

Also, with the new implementation of set_global_variable, it was no longer possible to supply values other than string as the second argument while the goal of this function is to support any value type. Since it is not supported by hcl2, we simply ignore this function for the hcl implementation, but we left it as it was for gotemplate.

* Remove warnings and unused stuff

* Change following comment

* Use the right return type on helper functions

* Complete the support of extended configuration files

Add support for various terragrunt configuration files (.hcl.json, .yaml, .config) to let users organize their repository as they want. It is now also possible to prefix the terragrunt configuration file with a dot to hide it and put it on the top of the IDE (separated from the terraform files)
Add a validation to determine if the folder pointed by source is containing terraform files (the previous version only validated this when no source was specified). It has been added as an option to not break existing code or tests, but this new behaviour is the default from the command line.
The cleanup now deletes all files that are containing terraform code (previously, only .tf files were deleted and .tf.json files were not). To achieve this, we no longer use the zglob package and rely on the utils.FindFiles with recursive option enabled.
Moved the config/FindConfigFilesInPath to options package where it is more convenient.
Added a function ConfigPath to find any valid configuration file within a path.
If a user specifies a --terragrunt-config argument, the name specified will also be used to resolve the stack (previously, that name was ignored for that command)
Added a detailed test to validate the various configurations options (including different configuration file format). The test also validates the various output of get-stack.
It is now possible to add a dependency to a parent folder even if that folder does not contain a terragrunt configuration file. Then, all terragrunt folders under that folder will be considered as a dependency.
Add the related documentation.
Update terraform to 0.12.26

* Revert naming to reduce the size of the PR

+ Fix a problem with conversion of non hcl config file
+ Fix matching non default config on simple command

* Apply suggestions from code review

Co-authored-by: ycanty <ycanty@coveo.com>

* Change following comments

+ Adjust a test that was failing sometimes and make its failing message clearer.

* Changes following review comments.

* Restore the DefaultConfigName refactoring

DefaultTerragruntConfigPath induces a false perception that this constant may be a path while it is actually a filename.

* Update gotemplate version

Fix a bug with sub template where error mode was automatically changed to strict.

Co-authored-by: ycanty <ycanty@coveo.com>
Adapt tests to new terraform extrapolation syntax