Change all includes to coveooss (#189)
The problem is since we are at version 2 and the upstream is still at 0.x, we cannot include our recent version of this module from another project. Event with replace, there is an incompatibility with the fact that the /v2 path is not specified.

If we replace with:
`replace github.com/gruntwork-io/terragrunt => github.com/coveooss/terragrunt v1.4.0`

it works. But if we try to update:
`replace github.com/gruntwork-io/terragrunt => github.com/coveooss/terragrunt v2.1.1`

It fails. We then have to specify:
`replace github.com/gruntwork-io/terragrunt => github.com/coveooss/terragrunt/v2 v2.1.1`

But it still fail because the module declaration in terragrunt was:
`module github.com/gruntwork.io/terragrunt`

while it must be:
`module github.com/coveooss/terragrunt/v2`

to make it work.

But then, all import statements in the terragrunt folder must be changed.