Improve error management
Since go 1.13, it is now possible to wrap errors by using fmt.Errorf with %w instead of %v. This allows better error management because we can check if the source of the error is of a specific type.

I used that to avoid useless delays when we try to load a non existing file from S3. Knowing that the root cause is not a network problem allows us to skip the multiple retries.