More flexible import variables
Currently, we use a terraform native function to load the default values of terraform variable in order to make them available to gotemplate. However, if the code contains gotemplate code, the odds that this code won't be initially compliant to terraform syntax are pretty high. This result in inability to load variables without using tricks such as GOVAR_xxx_ENDVAR.

If there are conditional part of code to select output or any other part of terraform code, the load default will also fail.

To fix that, we should only process the variables section of terraform code to read the variables default value (as it was before upgrading to terraform 0.12).