Change following review comments
It was more difficult than expected since the stderr stream was not redirected correctly. @jduchesne, do you see a problem with the change I did on args.go? The code you wrote had the effect of setting the streams on a copy of the logger stream and then assign them to the terragruntOptions object. Moreover, the stderr stream was set as the stdout stream on the logger. The SetOut and SetStdout are two distinct call. I was not able to evaluate if the debug message was issued correctly in the test without this change.

I added a test to evaluate the substitute on provider and avoid terraform complaining.

I added a test to confirm that we are able to load variables even if the terraform code in not valid (contains gotemplate generated name).