More flexible import variables (#221)
* Improve log information

There is not a big difference between the debug log level and the trace log level. To make it less verbose, I suggest to move some traces from debug to trace.

When we execute a sub-command, we may want to have a different logging level. I changed the code to be able to pass the desired log level.

Some trace in import_files were difficult to read. I simplified them to make them more readable.

There were some map initialization that do not require to specify type (resulting in lint warning). I removed them.

* More flexible import variables

Currently, we use a terraform native function to load the default values of terraform variable in order to make them available to gotemplate. However, if the code contains gotemplate code, the odds that this code won't be initially compliant to terraform syntax are pretty high. This result in inability to load variables without using tricks such as GOVAR_xxx_ENDVAR.

If there are conditional part of code to select output or any other part of terraform code, the load default will also fail.

To fix that, we should only process the variables section of terraform code to read the variables default value (as it was before upgrading to terraform 0.12).

* Change following review comment

* Change following review comments

It was more difficult than expected since the stderr stream was not redirected correctly. @jduchesne, do you see a problem with the change I did on args.go? The code you wrote had the effect of setting the streams on a copy of the logger stream and then assign them to the terragruntOptions object. Moreover, the stderr stream was set as the stdout stream on the logger. The SetOut and SetStdout are two distinct call. I was not able to evaluate if the debug message was issued correctly in the test without this change.

I added a test to evaluate the substitute on provider and avoid terraform complaining.

I added a test to confirm that we are able to load variables even if the terraform code in not valid (contains gotemplate generated name).

* Change provider type since there is no AWS configuration in the current test setup

* Add more control over default values cache