Fix tests
The provider template is deprecated. This provider has been used extensively in tests (both in the upstream Terragrunt and within our fork) because `data.template_file` was seen as a simple Terraform resource that didn’t produce any real artifacts. However, while the Linux versions are still available for download, the Windows and OSX/Darwin versions are no longer offered, causing test failures in local setups on Mac or Windows. This makes it challenging to add or modify tests locally.